requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
Flask==1.1.2